==> 3_calculate_shoe_distance.Rmd <==
---
title: "calculate_shoe_distance"
author: "Chelomo Lubliner"
date: "2023-09-21"
output: html_document
---


```{r }
library(dplyr)
ROOT_PATH = 'YOUR_PATH'
CONTOUR_ALGORITHM = 'Active_Contour'
# 'Convex'  or 'Active_Contour'
IMAGE_NUMBER  = 171
# 116 / 135
#  116/ 107 / 208/ 236 / 376 / 130 / 135

col_shoe<-307 #307 is the number of columns in each shoe
row_shoe<-395 #395 is the number of rows in each shoe
num_shoe<-387  #387 is the number of shoes but 386 is the number of shoes with RACs - shoe 127 has no RACS

rel_col_shoe<-150  #out of the 307 columns only 150 are relevant (contain non zero pixels in some shoes)
rel_row_shoe<-300  #out of the 395 rows only 300 are relevant (contain non zero pixels in some shoes)
rel_x_cord<-0.25 #using coordinates as in the locations_data.CSV file the relevant x coordinates are between -.25 and 0.25
rel_Y_cord<-0.5 #the relevant Y coordinates are between -0.5 and 0.5


file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Dataset/contour_algorithm.csv", sep = "")
contour_df <- read.csv(file_name,header=TRUE)


```




```{r }
calculate_distaxnce <- function(x1, y1, x2, y2) {
  sqrt((x1 - x2)^2 + (y1 - y2)^2)
}

min_distance_xy <- function(coordinates , contour_df, shoe_num){
  colnames(coordinates)[colnames(coordinates) == "Var1"] <- "x"
  colnames(coordinates)[colnames(coordinates) == "Var2"] <- "y"
  contour_df2 <- contour_df[contour_df$shoe ==shoe_num,]
  for (i in 1:nrow(coordinates)){
    percentage <- round(100 * i / nrow(coordinates),3)
    message("row N.",i, ": ", percentage, "%\n")
    min_distance <- Inf
    distance <- Inf
    for (j in 1:nrow(contour_df2)) {
      
        distance <- calculate_distance(coordinates$x[i], coordinates$y[i],contour_df2$x[j] , contour_df2$y[j]) 
        if (distance < min_distance) {
          min_distance <- distance
        }
      }
    
  coordinates$min_dist[i] <- min_distance
}
  file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Shoes_Distance/dist_shoe_",shoe_num,".csv", sep = "")
  write.csv(coordinates,file_name, row.names = FALSE)
  cat('file_saved : min_horiz_distance_xy')
  return(coordinates)
}

min_horiz_distance_xy <- function(coordinates , contour_df, shoe_num){
  colnames(coordinates)[colnames(coordinates) == "Var1"] <- "x"
  colnames(coordinates)[colnames(coordinates) == "Var2"] <- "y"
  contour_df1 <- contour_df[contour_df$shoe ==shoe_num,]
  for (i in 1:nrow(coordinates)){
    percentage <- round(100 * i / nrow(coordinates),3)
    message("row N.",i, ": ", percentage, "%\n")
    min_distance <- Inf
    distance <- Inf
    condition <- abs(coordinates$y[i] - contour_df1$y) < 0.01
    contour_df2 <- contour_df1[condition, ]
    if(nrow(contour_df2)> 1){
    for (j in 1:nrow(contour_df2)) {
      if (abs(coordinates$y[i] - contour_df2$y[j]) <0.01){
        distance <- calculate_distance(coordinates$x[i], coordinates$y[i],contour_df2$x[j] , contour_df2$y[j]) 
        if (distance < min_distance) {
          min_distance <- distance
        }
      }
    }
    }
  coordinates$horiz_dist[i] <- min_distance
  }
  coordinates$horiz_dist[is.infinite(coordinates$horiz_dist)] <- coordinates$min_dist[is.infinite(coordinates$horiz_dist)]
  file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Shoes_Distance/dist_shoe_",shoe_num,".csv", sep = "")
  write.csv(coordinates,file_name, row.names = FALSE)
  cat('file_saved : min_horiz_distance_xy')
  return(coordinates)
}
```

```{r }
categorical_distance <- function(coordinates, shoe_num){
#MIN DISTANCE
min_dist_mean <-   mean(coordinates$min_dist)
coordinates <- coordinates %>% mutate(min_dist_cat = ifelse(min_dist > 0.1, 1, 0))


# HORIZ DISTANCE
horiz_dist_mean <- mean(coordinates$horiz_dist)
coordinates <- coordinates %>% mutate(horiz_dist_cat = ifelse(horiz_dist > 0.1, 1, 0))

file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Shoes_Distance/dist_shoe_",shoe_num,".csv", sep = "")
  write.csv(coordinates,file_name, row.names = FALSE)
  cat('file_saved : min_horiz_distance_xy')
  return(coordinates)
}
```

```{r }
xcoor <- t(matrix(rep((-col_shoe/2+1:col_shoe)*rel_Y_cord/rel_col_shoe,row_shoe),col_shoe,row_shoe))
ycoor <- -matrix(rep((-row_shoe/2+1:row_shoe)*rel_Y_cord/rel_col_shoe,col_shoe),row_shoe,col_shoe)
xy <- expand.grid(xcoor[1,],ycoor[,1])#307 * 395 
xy <- min_distance_xy(xy,contour_df,IMAGE_NUMBER)

file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Shoes_Distance/dist_shoe_",IMAGE_NUMBER,".csv", sep = "")
xy <- read.csv(file_name,header=TRUE)
xy <- min_horiz_distance_xy(xy,contour_df,IMAGE_NUMBER)
xy <- categorical_distance(xy,IMAGE_NUMBER)

#rm(xy)
#rm(contour_df)
rm(xcoor)
rm(ycoor)
gc()
```
