==> 4_random_effect_model_adding_dist_as_covariate.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r }
set.seed(313)
# library(imager)
# library(spam)
# library(Matrix)
# library(lme4)
# library(splines)
# library(magrittr)
# library(imager)
# library(ggplot2) 
# library(fields)
library(imager)
library(ggplot2) 
library(plotly)
library(splines)
library(spam)
library(Matrix)
library(lme4)
library(rgl)
library(fields)
library(survival)
library(smoothie) 

col_shoe<-307 #307 is the number of columns in each shoe
row_shoe<-395 #395 is the number of rows in each shoe
num_shoe<-387  #387 is the number of shoes but 386 is the number of shoes with RACs - shoe 127 has no RACS

rel_col_shoe<-150  #out of the 307 columns only 150 are relevant (contain non zero pixels in some shoes)
rel_row_shoe<-300  #out of the 395 rows only 300 are relevant (contain non zero pixels in some shoes)
rel_x_cord<-0.25 #using coordinates as in the locations_data.CSV file the relevant x coordinates are between -.25 and 0.25
rel_Y_cord<-0.5 #the relevant Y coordinates are between -0.5 and 0.5
```

```{r }
ROOT_PATH = 'YOUR_PATH'
CLEANED_SHOES_PATH = 'YOUR_PATH/These/Files_thesis/Old_Shoes/Cleaned_Shoes/'

CONTOUR_ALGORITHM = 'Active_Contour'
# 'Convex' or 'Active_Contour' 
IMAGE_NUMBER  = '171'
# 171/ 135
#  117// 107 / 208/ 236 / 376 / allcont q 130

MODEL_FEATURE  = 'NS_XY_HORIZ_DUMMY'
# MODEL_FEATURE  = 'NULL_MODEL'
# MODEL_FEATURE  = 'NS_XY'
# MODEL_FEATURE  = 'NS_HORIZ'   -> 1.only Natural Cubic Splines with HORIZ
# MODEL_FEATURE  = 'NS_MIN'   -> 1. only Natural Cubic Splines with MIN DIST
# MODEL_FEATURE  = 'HORIZ_DIST' _> 2.only horiz distance 
# MODEL_FEATURE  = 'MIN_DIST' _> 2. only horiz distance 
# MODEL_FEATURE  = 'HORIZ_BIN_CAT'   -> 3.
# MODEL_FEATURE  = 'MIN_BIN_CAT' -> 3.
# MODEL_FEATURE  = 'NS_XY_HORIZ_DIST'
# MODEL_FEATURE  = 'NS_XY_MIN_DIST'
# MODEL_FEATURE  = 'NS_XY_HORIZ_BIN_CAT'
# MODEL_FEATURE  = 'NS_XY_MIN_BIN_CAT'
# MODEL_FEATURE  = 'NS_XY_NS_HORIZ'
# MODEL_FEATURE  = 'NS_XY_HORIZ_DUMMY'
# MODEL_FEATURE  = 'NS_XY_MIN_DUMMY'


```



```{r }


file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Dataset/dataCC_distance.csv", sep = "")
dataCC<- read.csv(file_name,header=TRUE)
file_name <- paste(ROOT_PATH,"Data/all_cont.csv", sep = "")
allcont<- read.csv(file_name,header=TRUE)

```






```{r }
#dataCC <-read.csv("C:/Users/Chelomo/Desktop/These/R/NEW_NAOMI_CHELOMO/SAVED_DATACC/convex_min_distance_21_09.csv",header=TRUE)
Random<-function(nknotsx=3,nknotsy=5,dat=dataCC,model_feat = MODEL_FEATURE )
{
  
  common_formula <- as.formula("n_Acc ~ (1 | shoe)")
  
  
  knotsx <- as.numeric(quantile(dat$x,1:nknotsx/(1+nknotsx)))
  knotsy <-as.numeric(quantile(dat$y,1:nknotsy/(1+nknotsy)))
  knots_distance <-as.numeric(quantile(dat$horiz_dist,1:2/(3)))
  shoe<-dat$shoe
  
  
  if(model_feat == 'NS_XY'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + (1 | shoe) "}
  else if(model_feat == 'NS_MIN'){
            formula_text <- "n_Acc ~ ns(dat$min_dist,knots=knots_distance) + (1 | shoe) "}
  else if(model_feat == 'EMPTY_MODEL'){
            formula_text <- "n_Acc ~ (1 | shoe) "}
  else if(model_feat == 'NS_HORIZ'){
            formula_text <- "n_Acc ~ ns(dat$horiz_dist,knots=knots_distance) + (1 | shoe) "}
  else if(model_feat == 'HORIZ_DIST'){
            formula_text <- "n_Acc ~ dat$horiz_dist + (1 | shoe)"}
  else if(model_feat == 'MIN_DIST'){
            formula_text <- "n_Acc ~ dat$min_dist + (1 | shoe)"}
  else if(model_feat == 'HORIZ_BIN_CAT'){
            formula_text <- "n_Acc ~ dat$horiz_dist_cat + (1 | shoe)"}
  else if(model_feat == 'MIN_BIN_CAT'){
            formula_text <- "n_Acc ~ dat$min_dist_cat + (1 | shoe)"}
  else if(model_feat == 'NS_XY_HORIZ_DIST'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$horiz_dist + (1 | shoe) "}
  else if(model_feat == 'NS_XY_MIN_DIST'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$min_dist + (1 | shoe) "}
  else if(model_feat == 'NS_XY_HORIZ_BIN_CAT'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$horiz_dist_cat + (1 | shoe) "}
    else if(model_feat == 'NS_XY_HORIZ_DUMMY'){
      dat$dummy_0to1 <- as.numeric(dat$horiz_dist <= 0.05)
      dat$dummy_1to2 <- as.numeric(dat$horiz_dist > 0.05 & dataCC$horiz_dist <= 0.1)
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$dummy_0to1 + dat$dummy_1to2 +  (1 | shoe) "
    }
  else if(model_feat == 'NS_XY_MIN_DUMMY'){
      dat$dummy_0to1 <- as.numeric(dat$min_dist <= 0.05)
      dat$dummy_1to2 <- as.numeric(dat$min_dist > 0.05 & dataCC$min_dist <= 0.1)
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$dummy_0to1 + dat$dummy_1to2 +  (1 | shoe) "
            }
  else if(model_feat == 'NS_XY_MIN_BIN_CAT'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + dat$min_dist_cat + (1 | shoe) "}
    else if(model_feat == 'NS_XY_NS_HORIZ'){
            formula_text <- "n_Acc ~ ns(dat$x,knots=knotsx):ns(dat$y,knots=knotsy) + ns(dat$horiz_dist,knots=knots_distance) + (1 | shoe) "}
  else { stop("Invalid model_feat value")}
  cat(formula_text)
  est_formula <- as.formula(formula_text)  
  initial_values <- c(parameter1 = 0)  # Adjust these values
  est<- glmer(est_formula, data = dat, family = binomial(link = "logit"), control =glmerControl(optimizer = "bobyqa"))
  #"Nelder_Mead", optCtrl = list(maxfun=2e5)
  cat("file_saving")
  file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Saved_Models/", model_feat,".rds", sep = "")
  saveRDS(est, file = file_name)
  return(est)
}
rand<-Random(dat=dataCC)

```



```{r }

file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Saved_Models/", MODEL_FEATURE,".rds", sep = "")
rand <- readRDS(file = file_name)
#print(rand)
#print(summary(rand))
xcoor <- t(matrix(rep((-col_shoe/2+1:col_shoe)*rel_Y_cord/rel_col_shoe,row_shoe),col_shoe,row_shoe))
ycoor <- -matrix(rep((-row_shoe/2+1:row_shoe)*rel_Y_cord/rel_col_shoe,col_shoe),row_shoe,col_shoe)

#IN python it's starts from 0, so you have to add one value in the folder.
# since shoe n.126, it's the same value
if (IMAGE_NUMBER == 'allcont'){
  img <- allcont
  xcoor <- t(matrix(rep((-col_shoe/2+1:col_shoe)*rel_Y_cord/rel_col_shoe,row_shoe),col_shoe,row_shoe))
  ycoor <- -matrix(rep((-row_shoe/2+1:row_shoe)*rel_Y_cord/rel_col_shoe,col_shoe),row_shoe,col_shoe)
  xy <- expand.grid(xcoor[1,],ycoor[,1])#307 * 395  
  colnames(xy)[colnames(xy) == "Var1"] <- "x"
  colnames(xy)[colnames(xy) == "Var2"] <- "y"
}else{
  # since this file contains the shoe without RAC (126), that has been removed. and python files start at 0 
  # we will remove 1 to the cleaned_shoe if it's < 127
  cleaned_im_num <- if (as.integer(IMAGE_NUMBER) < 127) as.integer(IMAGE_NUMBER) -1 else IMAGE_NUMBER  
  img <- load.image(paste( CLEANED_SHOES_PATH, "im_",cleaned_im_num, ".png", sep = ""))
  img <- t(as.matrix(img))
  file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Shoes_Distance/dist_shoe_",IMAGE_NUMBER,".csv", sep = "")
  ## PROBLEM!!
  #xy <- expand.grid(xcoor[1,],ycoor[,1])#307 * 395
  #colnames(xy)[colnames(xy) == "Var1"] <- "x"
  #colnames(xy)[colnames(xy) == "Var2"] <- "y"
  xy<- read.csv(file_name, header=TRUE)
  #read.csv("C:/Users/Chelomo/Desktop/These/R/NEW_NAOMI_CHELOMO/SAVED_DATACC/convex_min_horiz_distance_xy_107.csv",header=TRUE)
}


```


```{r }


newdesignmat <- rep(1,length(xy$x))
#we multiplicate the splines(x) with splines(y) so newdesigmat have 4*6 columns! (plus the first)
#newdesgmat have 12165 lines, exactly like ROWS*COLS. it will represent each point of the matrix ?


if (grepl("NS_XY", MODEL_FEATURE)){ # Splines 
  nknotsx <- 3
  nknotsy <- 5
  knotsx <- as.numeric(quantile(dataCC$x,1:nknotsx/(1+nknotsx)))
  knotsy <-as.numeric(quantile(dataCC$y,1:nknotsy/(1+nknotsy)))
  basx <- ns(dataCC$x,knots=knotsx)
  basy <- ns(dataCC$y,knots=knotsy)
  for(i in 1:length(predict(basy,1))) {
    for(j in 1:length(predict(basx,1))) {
      newdesignmat <-  cbind(newdesignmat,predict(basx, xy$x)[,j]*predict(basy, xy$y)[,i])  } }
} 
if (grepl("NS_HORIZ", MODEL_FEATURE)){ 
   knots_distance <-as.numeric(quantile(dataCC$horiz_dist,1:2/(3)))
  bas_horiz <- ns(dataCC$horiz_dist,knots=knots_distance)
  newdesignmat <- cbind(newdesignmat,predict(bas_horiz,xy$horiz_dist))
} else if (grepl("NS_MIN", MODEL_FEATURE)){ 
  knots_distance <-as.numeric(quantile(dataCC$min_dist,1:2/(3)))
  bas_min <- ns(dataCC$min_dist,knots=knots_distance)
  newdesignmat <- cbind(newdesignmat,predict(bas_min, xy$min_dist))
} else if(grepl("HORIZ_DUMMY", MODEL_FEATURE)){
  xy$dummy_0to1 <- as.numeric(xy$horiz_dist <= 0.05)
  xy$dummy_1to2 <- as.numeric(xy$horiz_dist > 0.05 & xy$horiz_dist <= 0.1)
  newdesignmat <- cbind(newdesignmat,xy$dummy_0to1)
  newdesignmat <- cbind(newdesignmat,xy$dummy_1to2)
} else if(grepl("MIN_DUMMY", MODEL_FEATURE)){
  xy$dummy_0to1 <- as.numeric(xy$min_dist <= 0.05)
  xy$dummy_1to2 <- as.numeric(xy$min_dist > 0.05 & xy$min_dist <= 0.1)
  newdesignmat <- cbind(newdesignmat,xy$dummy_0to1)
  newdesignmat <- cbind(newdesignmat,xy$dummy_1to2)
}



#variables
if (grepl("HORIZ_DIST", MODEL_FEATURE)){
  newdesignmat <- cbind(newdesignmat,xy$horiz_dist)
}else if (grepl("MIN_DIST", MODEL_FEATURE)){
  newdesignmat <- cbind(newdesignmat,xy$min_dist)
}else if (grepl("HORIZ_BIN_CAT", MODEL_FEATURE)){
  newdesignmat <- cbind(newdesignmat,xy$horiz_dist_cat )
}else if (grepl("MIN_BIN_CAT", MODEL_FEATURE)){
  newdesignmat <- cbind(newdesignmat,xy$min_dist_cat)
}



# IMAGES of the model

#(for each row, we multiply and sum the params between the two matrices )
#shape of matrix multiplication: (121265xN.Params)*(N.Params x 1) -> (121265x1)
pred.case_control <- newdesignmat%*%fixef(rand)+log(0.005) #log(0.005) is the offset
#CHANGE THE NEXT LINE

#in my model -> remove only pixels that out of the SPECIFIC shoe

#allcont -> img
pred.case_control[t(img)==0] <- NA #areas out of the contour (less than 8 shoes has contact surface in these pixels) are given NA
prob.pred <- exp(matrix(pred.case_control ,row_shoe,col_shoe,byrow=1))/(1+exp(matrix(pred.case_control ,row_shoe,col_shoe,byrow=1)))
intens <- -log(1-prob.pred) #turning it to intensity
rotated_intens <- t(intens)[, nrow(intens):1]
#pdf("C:/Users/Chelomo/Desktop/These/R/NEW_NAOMI_CHELOMO/SAVED_IMAGES/random_effect.pdf")
image.plot(rotated_intens, axes = TRUE)
#dev.off()

```






```{r}
file_name <- paste(ROOT_PATH, CONTOUR_ALGORITHM, "/Model_Images/",MODEL_FEATURE, "_",IMAGE_NUMBER, ".pdf", sep = "")
pdf(file = file_name, height = 6, width = 6)
image_dims <- dim(rotated_intens)
image_length <- image_dims[1]
image_width <- image_dims[2]
image.plot(rotated_intens, axes = TRUE)

abline(v = 0.5, col = "black", lty = 1, lwd = 1)
abline(h = 0.5, col = "black", lty = 1, lwd = 1)
grid(nx = 10, ny = 10, col = "gray30", lty = "dotted")
dev.off()

```

```{r}

```
